better naming / cleaner handling of set patterns
