+++ lib/cardlib/permissions.rb
@@ -264,7 +264,7 @@ module Cardlib::Permissions
         in_set = {}
         if !(self.trash)
           if class_id = (set=left and set_class=set.tag and set_class.id)
-            rule_class_ids = Cardlib::Pattern.subclasses.map &:key_id
+            rule_class_ids = set_patterns.map &:key_id
             #warn "rule_class_id #{class_id}, #{rule_class_ids.inspect}"
 
             #first update all cards in set that aren't governed by narrower rule
